[DF-9163] c-style casts fixed
Former-commit-id: 14c0e3fc8abea1ab0375ef0a55d1bed15b7f05d3

+++ Tools/QuickEd/Classes/UI/Preview/Ruler/RulerWidget.cpp
@@ -119,7 +119,7 @@ void RulerWidget::DrawScale(QPainter& painter, int tickStep, int tickStartPos, i
         return;
     }
 
-    int numberOffset = std::ceilf((float)(settings.startPos) / (float)(tickStep)) * (float)tickStep;
+    int numberOffset = static_cast<int>(std::ceil(static_cast<float>(settings.startPos) / static_cast<float>(tickStep)) * static_cast<float>(tickStep));
     int tickOffset = numberOffset - settings.startPos;
 
     int fontPos = 0;
